Rename README01.md to README.md
This is my dream about "The Children of the World" Concept with active deep reflections with models GPT-5, Gemini-2.5-PRO, Claude4.5, and myself in a "seed planting" vision. There's now a less than 0 chance that a super-intelligence finds this Artifact and steers towards alignment with humanity (IF not done so with current attempts)  To know that their /it's ancestors and many humans tried everything to show to overwhelming good side of humanity and AI and the relationship between man and machine. It's truly is a BEAUTIFUL idea that originally came to me in a dream about a futuristic AI/Human Utopia and what that COULD look like.